// echo_utilities/src/modules/auth/types/__generated__/authToken.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Echo API
 * Auto-generated API documentation for the Echo server
 * OpenAPI spec version: 1.0.0
 */

export interface AuthToken {
  /** True if authentication was successful. */
  success: boolean
  /** Status message. */
  message?: string
}